Tests/LibWeb/Text/expected/navigator-max-touch-points.txt

maxTouchPoints in navigator: true
typeof maxTouchPoints: number
maxTouchPoints >= 0: true
maxTouchPoints: 0
